chore: Rename component and update comment
