+++ components/sm_gauge.py
